django_dbchat/requirements.txt
# Django and web framework requirements
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-redis==5.4.0
channels==4.0.0
channels-redis==4.1.0
django-extensions==3.2.3

# Database drivers
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
# mysqlclient==2.2.0  # Commented out for Docker compatibility
# oracledb==1.4.2     # Commented out for Docker compatibility  
# pyodbc==5.0.1       # Commented out for Docker compatibility

# Data processing and analysis
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# Visualization
plotly==5.17.0
kaleido==0.2.1

# LLM and AI
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2

# Background tasks
celery==5.3.4
redis==5.0.1

# Email
django-email-extras==0.3.4

# File handling and utilities
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.4.0

# PDF generation
weasyprint==62.3
reportlab==4.2.2

# Authentication and security
django-allauth==0.57.0
cryptography==42.0.8

# API documentation
drf-spectacular==0.26.5

# Development and testing
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Monitoring and logging
sentry-sdk==1.38.0
django-health-check==3.17.0

# Additional utilities
requests==2.31.0
urllib3==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# Added from the code block
mysql-connector-python==8.2.0
xlrd==2.0.1

# PDF and Image Generation
selenium==4.15.2 